requirements.txt
python-telegram-bot==13.7
urllib3==1.26.6
requests==2.26.0
pillow==8.4.0